Initial Commit
Will be adding more sensors support soon. Waiting on them to be shipped.
So far there is the DHT11 Humidity and Temerature sensor support. Writes
to a MySQL table named data in a database called weather_data. Also
saves an image from the camera with the current Temp (C & F) along with
humidity. Saves the image as Image_#.jpg with a quality of 100.
Also has an ini config file (settings.ini)