some small changes
changed some defaults, added some comments, and some strings.
My dht11 humidity sensor is returning a value that I do not think is
right, if it is doubled, it is close to correct. Not sure what is going
on. Have not read any articles or anything about the value that is
returned from the dht11 python library needing to be multiplied by 2.
The temperature is returning a correct value, so that part of it is
working.