some changes
Fixed the issue where the settings address was not being converted to an
int
Changed photoresistor_pin to photoresistor_channel
Changed the setup for the PFC and BMP.
Fixed the re-setting up the PCF for in the BMP setup method
Renamed dht_pin to dht11_pin and dht22_pin so support dht22 in the
future
Added BMP flag support to define between the BMP085 (current support),
180 and 280 (future support)
Added support to enable camera if you would like it.
Added SQL logging of pressure and photolevel data.
Humidity is in the sensors raw value that is read. for the DHT11 it
needs to be doubled, not sure about the DHT22 yet.