More Front End work
Got the Graphs made for the station view. Humidity and pressure are
separated into their own graphs from the temperature data of each sensor
so that it is not rendered as a flat line.
Graphs are the newest 100 points of data for now, might add an option to
choose the amount of data to view.

Also removed station_sensors.tpl and PageCount.tpl as they are not
needed any more