More work and half done code...
Added the Last Update time also to the daemon since it wasnt doing that
before.

Graph now only shows time, camera image is stil being worked on.

More code for the am2302, still not working though, data is not being
read for some reason.

Added in filtering of 0 data from the queries on the front end. makes
for a cleaner graph

Starting work on Front end API to take data from stations as an
alternate to an SQL connection in the Python Daemon.