Fixed a few things that got broken
Index page was not longer getting its data the way it was expecting. So
instead of making a third function to get sensor data, i condensed them
down into one. Limit defines what is returned. 0 for all rows, 1 for the
newest dataset, any other number and it will return that many rows.

also fixed the photoresistor graph not showing the time, also did a temp
implemetation of converting the time to the local time of the browsers
choosing since the data is stored as GMT... or is it UTC.... I forgot...
It should be in the code for the daemon.