Moved a bunch of things around and added a few things
Working on the API upload to the HTML front end. /piwem/api/

Moved the sensor value data structure to be more of a package and have
sub-objects for each sensor.
Also created a payload object that gets converted to a JSON string to be
sent to the API for logging.

Need to make a better way of registering a new Station to the API. Will
probably change it to be a form with a login on the HTML front end.