Added support for local buffering of data
If the central upload server is not available, then you can choose to
buffer the data locally until it is avaiable again. Once it is
available, the script will upload the data and then clean it out.
This will not touch data that was logged when choosing the localstorage
option. That option is separate from the bufferlocally option.