bug fixes
timstamp from buffered uploads to the central server were using the
current time at upload and not the timestsamp from the table.

removed set_sensor_flags function and moved it to the init.

sample settings was missing some values. also added in the ones that are
needed so far for the anemometer and wind vane.