Some small changes
Added in the INA3221 library to support the 3 channel Voltage/Current
monitor. Modified the power.py script to use it. Toggle the use_ina3221
var to 1 to use INA3221, or set it to 0 to use the INA219 single
channel.

Updated the blank.sql to include the channel if using the 3221.

removed a mention of WifiDB from the readme.md that it was copied from.