Massive Update
Added in the Power module to monitor power usage of the Weather Stations various power supplies and drains.

Added skeleton of a Client to connect to the locally run web server on the weather station to view its current data. Also will have it be able to connect to a central server to be able to view the data for many weather stations.

Added in the wind module, but need to work on how the anemometer clicks of the magnetic relay are detected. It seems to not get the faster (more than 5mph :disappointed: )

A few more small enhancements